# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat Log API",
    "description": "记录聊天信息的API服务",
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "开发服务器"
    }
  ],
  "paths": {
    "/logs": {
      "get": {
        "summary": "获取聊天日志",
        "description": "获取指定日期范围内的聊天日志",
        "operationId": "getLogs",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "开始日期 (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "结束日期 (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取日志",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": ["logs"]
      },
      "post": {
        "summary": "创建聊天日志",
        "description": "创建一条新的聊天日志记录",
        "operationId": "createLog",
        "requestBody": {
          "description": "日志信息",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLogRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建日志",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": ["logs"]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateLogRequest": {
        "type": "object",
        "description": "创建日志的请求参数",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string",
            "description": "聊天文本内容",
            "example": "这是一条聊天记录"
          },
          "validation_result": {
            "type": "boolean",
            "description": "验证结果",
            "example": true
          }
        }
      },
      "LogResponse": {
        "type": "object",
        "description": "单条日志操作的响应结果",
        "properties": {
          "code": {
            "type": "integer",
            "description": "操作状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "操作消息",
            "example": "日志保存成功"
          },
          "data": {
            "type": "object",
            "description": "日志数据",
            "properties": {
              "id": {
                "type": "string",
                "description": "日志唯一标识符",
                "example": "202503031215ABcd12"
              },
              "text": {
                "type": "string",
                "description": "聊天文本内容",
                "example": "这是一条聊天记录"
              },
              "validation_result": {
                "type": "boolean",
                "description": "验证结果",
                "example": true
              }
            }
          }
        }
      },
      "LogsResponse": {
        "type": "object",
        "description": "多条日志查询的响应结果",
        "properties": {
          "code": {
            "type": "integer",
            "description": "操作状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "操作消息",
            "example": "成功获取日志"
          },
          "data": {
            "type": "array",
            "description": "日志数据列表",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "日志唯一标识符",
                  "example": "202503031215ABcd12"
                },
                "text": {
                  "type": "string",
                  "description": "聊天文本内容",
                  "example": "这是一条聊天记录"
                },
                "validation_result": {
                  "type": "boolean",
                  "description": "验证结果",
                  "example": true
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "错误响应结果",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误状态码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误消息",
            "example": "无效的请求参数"
          }
        }
      }
    }
  }
}